IssueNo: #I5KB49 Description: 两次删除同一个form都返回成功修复 Sig: SIG_ApplicationFramework Feature or Bugfix: Bugfix Binary Source: No

+++ services/src/form_mgr_adapter.cpp
@@ -311,7 +311,10 @@ ErrCode FormMgrAdapter::HandleDeleteTempForm(const int64_t formId, const sptr<IR
             FormDataMgr::GetInstance().AddFormUserUid(formId, uid);
             return result;
         }
-        FormDataMgr::GetInstance().DeleteTempForm(formId);
+        if (!FormDataMgr::GetInstance().DeleteTempForm(formId)) {
+            HILOG_ERROR("%{public}s, form id is not existed.", __func__);
+            return ERR_APPEXECFWK_FORM_NOT_EXIST_ID;
+        }
         FormDataMgr::GetInstance().DeleteFormRecord(formId);
         if (!FormCacheMgr::GetInstance().DeleteData(formId)) {
             HILOG_ERROR("%{public}s, failed to remove cache data", __func__);
@@ -352,7 +355,7 @@ ErrCode FormMgrAdapter::HandleDeleteFormCache(FormRecord &dbRecord, const int ui
         }
         if (!FormDataMgr::GetInstance().DeleteFormRecord(formId)) {
             HILOG_ERROR("%{public}s, failed to remove cache data", __func__);
-            return ERR_APPEXECFWK_FORM_COMMON_CODE;
+            return ERR_APPEXECFWK_FORM_NOT_EXIST_ID;
         }
         if (result = FormDbCache::GetInstance().DeleteFormInfo(formId); result != ERR_OK) {
             HILOG_ERROR("%{public}s, failed to remove db data", __func__);
